requirements.txt
streamlit>=1.38
pandas>=2.2
openpyxl>=3.1
plotly>=5.24
google-genai>=0.8
python-dotenv>=1.0
